requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
Flask-Compress==1.14
psycopg2-binary==2.9.7
cloudinary==1.36.0
python-dotenv==1.0.0
pytz==2024.1
bcrypt==4.0.1
WTForms==3.0.1
requests==2.31.0
gunicorn==21.2.0
email-validator==2.1.0
greenlet==2.0.2
SQLAlchemy==2.0.23
blinker==1.6.3
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
Pillow==10.0.0
groq==0.3.0